Shell Firing
Implemented shell firing with the space button, done through the TankShooting script.
The script calculates the charging force of the shots (by holding down Space) and translate the force to the fired shell's velocity, then do the Fire() function that instantiate the shell object and move it accordingly.
Script also updates the AimSlider slider UI