Synchronized TankHealth
Synchronized the TankHealth script so that the tank explosion happens in Clients as well.
This includes: currentHealth, dead variables
as well as changed how the OnDeath (now OnDeathServerRpc) function works: destroy the tank istead of just disabling it, and spawning in the explosion particle effect (with the prefab changed to play on awake) in the same time of the tank's destruction instead of spawning it beforehand and playing it then.